# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "Complete E-Commerce API with Xendit Payment Integration and Automated Webhook Processing",
        "contact": {
            "name": "API Support",
            "email": "support@ecommerce.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ecommerce-api-test-production.up.railway.app",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new order",
                "operationId": "c90fce9fa5e9860be3ce8ede472afc4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "shipping_address"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "maximum": 10,
                                                    "minimum": 1,
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "shipping_address": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "minLength": 10,
                                        "example": "Jl. Contoh Alamat No. 123, Jakarta Selatan, 12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order berhasil dibuat"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "user_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "order_number": {
                                                            "type": "string",
                                                            "example": "ORD-1705312800-1234"
                                                        },
                                                        "total_amount": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 30000000
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "pending"
                                                        },
                                                        "shipping_address": {
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "product_id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "quantity": {
                                                                        "type": "integer"
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "format": "float"
                                                                    },
                                                                    "subtotal": {
                                                                        "type": "number",
                                                                        "format": "float"
                                                                    },
                                                                    "product": {
                                                                        "type": "object"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "payment_url": {
                                                    "type": "string",
                                                    "example": "https://your-app.railway.app/api/payments/create/1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Checkout failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user's order history",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar order berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "order_number": {
                                                                "type": "string"
                                                            },
                                                            "total_amount": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "pending",
                                                                    "paid",
                                                                    "processing",
                                                                    "shipped",
                                                                    "delivered",
                                                                    "cancelled"
                                                                ]
                                                            },
                                                            "shipping_address": {
                                                                "type": "string"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "items": {
                                                                "type": "array",
                                                                "items": {}
                                                            },
                                                            "payment": {
                                                                "type": "object",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail order berhasil diambil"
                                        },
                                        "data": {
                                            "properties": {
                                                "order": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "order_number": {
                                                            "type": "string"
                                                        },
                                                        "total_amount": {
                                                            "type": "number",
                                                            "format": "float"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        },
                                                        "shipping_address": {
                                                            "type": "string"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {}
                                                        },
                                                        "payment": {
                                                            "type": "object",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payments/create/{orderId}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create payment invoice",
                "operationId": "b506c5153692b98ba632b0be60301e57",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment invoice created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Payment invoice created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "payment_url": {
                                                    "type": "string",
                                                    "example": "https://checkout.xendit.co/web/abc123"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthorized access to order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API Access Key",
                "name": "X-ACCESS-KEY",
                "in": "header"
            },
            "secretKey": {
                "type": "apiKey",
                "description": "API Secret Key",
                "name": "X-SECRET-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}